+++ RBX-Audio-Extractor2/obj/Debug/net10.0-windows/win-x64/MainWindow.g.i.vb
@@ -188,7 +188,7 @@ Partial Public Class MainWindow
             Return
         End If
         _contentLoaded = true
-        Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.1.0;component/mainwindow.xaml", System.UriKind.Relative)
+        Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.2.0;component/mainwindow.xaml", System.UriKind.Relative)
         
         #ExternalSource("..\..\..\..\MainWindow.xaml",1)
         System.Windows.Application.LoadComponent(Me, resourceLocater)