+++ RBX-Audio-Extractor2/obj/Debug/net10.0-windows/win-x64/RBX-Audio-Extractor2.AssemblyInfo.vb
@@ -15,11 +15,11 @@ Imports System
 Imports System.Reflection
 <Assembly: System.Reflection.AssemblyCompanyAttribute("RBXAssetExtractor"),  _
  Assembly: System.Reflection.AssemblyConfigurationAttribute("Debug"),  _
- Assembly: System.Reflection.AssemblyFileVersionAttribute("3.1.1.0"),  _
- Assembly: System.Reflection.AssemblyInformationalVersionAttribute("3.1.1+7b34bbc7d9630efd4bc87eef9bc888900144fe8d"),  _
+ Assembly: System.Reflection.AssemblyFileVersionAttribute("3.1.2.0"),  _
+ Assembly: System.Reflection.AssemblyInformationalVersionAttribute("3.1.2+b8df7b65a5abe3768ee8f16adc2c9800412e77c5"),  _
  Assembly: System.Reflection.AssemblyProductAttribute("RBXAssetExtractor"),  _
  Assembly: System.Reflection.AssemblyTitleAttribute("RBXAssetExtractor"),  _
- Assembly: System.Reflection.AssemblyVersionAttribute("3.1.1.0"),  _
+ Assembly: System.Reflection.AssemblyVersionAttribute("3.1.2.0"),  _
  Assembly: System.Runtime.Versioning.TargetPlatformAttribute("Windows7.0"),  _
  Assembly: System.Runtime.Versioning.SupportedOSPlatformAttribute("Windows7.0")>