+++ RBX-Audio-Extractor2/obj/Debug/net10.0-windows/win-x64/Views/AudioView.g.i.vb
@@ -175,7 +175,7 @@ Namespace Views
                 Return
             End If
             _contentLoaded = true
-            Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.1.0;component/views/audioview.xaml", System.UriKind.Relative)
+            Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.2.0;component/views/audioview.xaml", System.UriKind.Relative)
             
             #ExternalSource("..\..\..\..\..\Views\AudioView.xaml",1)
             System.Windows.Application.LoadComponent(Me, resourceLocater)