+++ RBX-Audio-Extractor2/obj/Debug/net10.0-windows/win-x64/Views/MaintenanceView.g.i.vb
@@ -112,7 +112,7 @@ Namespace Views
                 Return
             End If
             _contentLoaded = true
-            Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.1.0;component/views/maintenanceview.xaml", System.UriKind.Relative)
+            Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.2.0;component/views/maintenanceview.xaml", System.UriKind.Relative)
             
             #ExternalSource("..\..\..\..\..\Views\MaintenanceView.xaml",1)
             System.Windows.Application.LoadComponent(Me, resourceLocater)