+++ RBX-Audio-Extractor2/obj/Release/net10.0-windows/win-x64/DialogWindow.g.i.vb
@@ -90,7 +90,7 @@ Partial Public Class DialogWindow
             Return
         End If
         _contentLoaded = true
-        Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.1.0;component/dialogwindow.xaml", System.UriKind.Relative)
+        Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.2.0;component/dialogwindow.xaml", System.UriKind.Relative)
         
         #ExternalSource("..\..\..\..\DialogWindow.xaml",1)
         System.Windows.Application.LoadComponent(Me, resourceLocater)