+++ RBX-Audio-Extractor2/obj/Release/net10.0-windows/win-x64/FontPreviewWindow.g.i.vb
@@ -111,7 +111,7 @@ Partial Public Class FontPreviewWindow
             Return
         End If
         _contentLoaded = true
-        Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.1.0;component/fontpreviewwindow.xaml", System.UriKind.Relative)
+        Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.2.0;component/fontpreviewwindow.xaml", System.UriKind.Relative)
         
         #ExternalSource("..\..\..\..\FontPreviewWindow.xaml",1)
         System.Windows.Application.LoadComponent(Me, resourceLocater)