+++ RBX-Audio-Extractor2/obj/Release/net10.0-windows/win-x64/Views/AboutView.g.i.vb
@@ -105,7 +105,7 @@ Namespace Views
                 Return
             End If
             _contentLoaded = true
-            Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;component/views/aboutview.xaml", System.UriKind.Relative)
+            Dim resourceLocater As System.Uri = New System.Uri("/RBXAssetExtractor;V3.1.2.0;component/views/aboutview.xaml", System.UriKind.Relative)
             
             #ExternalSource("..\..\..\..\..\Views\AboutView.xaml",1)
             System.Windows.Application.LoadComponent(Me, resourceLocater)